Issue #25: Introduce concrete ListParameter and SimpleParameter
Introduce ConfigurationListParameter and ConfigurationSimpleParameters
as ConfigurationParameter derivatives. This once again changes the
architecture from one where ConfigurationParameter is a concrete class
implementing the "simple" functionality to one where it is an abstract
class, and the "simple" (and "list") functionality is implemented in
derived concrete classes.  It's not architecturally possible for
ConfigurationParameter to contain the simple parameter content and
simultaneously have the ListParameter class use its functionality.

Partially implement ConfigurationListParamter_test.

TemplateClass is updated with new 110 line wrap limit.