Issue #25: Fully implement ConfigurationSimpleParameter_test
Implement ConfigurationSimpleParameter_test.  All current
ConfigurationSimpleParameter content should be covered.  The fromString
and toString tests content came from ConfigurationParameter_test, which
is where it was when ConfigurationSimpleParameter was implemented as
part of ConfigurationParameter.

ConfigurationParameter itself is abstract, so a
ConfigurationNoopParameter class is introduced specifically for this
test to provide a concrete implementation that can be tested.  It's
called "noop" because its implementation for the abstract
ConfigurationParameter functionality does nothing. Concrete examples of
useful implementations of the abstract ConfigurationParameter
functionality should be tested elsewhere.